+++ packages/utils.jsx/src/index.js
@@ -1,2 +1,7 @@
-export {getOriginalJsxForNode, maybeJsx} from './jsx'
-export {default as JsxObserver} from './JsxObserver'
+export {
+  getOriginalJsxForNode, maybeJsx, createElement
+} from './jsx'
+
+export {
+  default as JsxObserver
+} from './JsxObserver'