+++ packages/utils.jsx/src/jsx.js
@@ -47,3 +47,18 @@ export function maybeJsx (possibleJsx) {
 export function getOriginalJsxForNode (node) {
   return node[ORIGINAL_JSX_SYM]
 }
+
+/**
+ * Convert JSX into an object that can be consumed by TKO.
+ * Mimics React.createElement
+ * @param {string} e tagName of the element
+ * @param {object|null} a attributes of the element
+ * @param  {...string|object} c children of the element
+ */
+export function createElement (e, a, ...c) {
+  return {
+    elementName: e,
+    attributes: a || {},
+    children: [...c]
+  }
+}