+++ CHANGELOG.md
@@ -3,6 +3,7 @@ For TODO between alpha and release, see https://github.com/knockout/tko/issues/1
 ## 🎇 alpha-7.x
 
 * Make node cleaning upon removal with JSX asynchronous
+* Add `createElement` to the `tko` reference build, that (essentially) mimics `React.createElement`
 
 ## 🎒 `alpha-5a..alpha6.x`  (ongoing)